requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
beanie==1.26.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
